Wip release

+++ test-all.sh
@@ -5,8 +5,8 @@
 # Used by Jenkins to verify commits.
 
 # add new DSLs to test here
-dsls=( "SimpleVector" "OptiML" )
-runners=( "ppl.dsl.forge.examples.SimpleVectorDSLRunner" "ppl.dsl.forge.dsls.optiml.OptiMLDSLRunner" )
+dsls=( "SimpleVector" "OptiML" "OptiQL" "OptiGraph" "OptiWrangler" )
+runners=( "ppl.dsl.forge.examples.SimpleVectorDSLRunner" "ppl.dsl.forge.dsls.optiml.OptiMLDSLRunner" "ppl.dsl.forge.dsls.optiql.OptiQLDSLRunner" "ppl.dsl.forge.dsls.optigraph.OptiGraphDSLRunner" "ppl.dsl.forge.dsls.optiwrangler.OptiWranglerDSLRunner" )
 
 # exit if any part of the script fails
 set -e
@@ -19,6 +19,12 @@ if [ -z "${LMS_HOME}" ]; then echo error: LMS_HOME is not defined; exit $E_BADEN
 if [ -z "${DELITE_HOME}" ]; then echo error: DELITE_HOME is not defined; exit $E_BADENV; fi
 if [ -z "${FORGE_HOME}" ]; then echo error: FORGE_HOME is not defined; exit $E_BADENV; fi
 
+# check for required configuration files
+if [ ! -f "${DELITE_HOME}/config/delite/CPP.xml" ]; then echo error: CPP.xml is not present; exit $E_BADENV; fi
+if [ ! -f "${DELITE_HOME}/config/delite/BLAS.xml" ]; then echo error: BLAS.xml is not present; exit $E_BADENV; fi
+if [ ! -f "${DELITE_HOME}/config/delite/CUDA.xml" ]; then echo error: CUDA.xml is not present; exit $E_BADENV; fi
+if [ ! -f "${DELITE_HOME}/config/delite/cuBLAS.xml" ]; then echo error: cuBLAS.xml is not present; exit $E_BADENV; fi
+
 # remove previous delite runtime cache
 rm -rf $DELITE_HOME/generatedCache
 
@@ -30,6 +36,10 @@ sbt -Dtests.threads=1 -Dtests.targets=scala,cpp "; project tests; test"
 echo "[test-all]: running Delite and Delite DSL tests (8 threads)"
 sbt -Dtests.threads=8 -Dtests.targets=scala,cpp "; project tests; test"
 
+# delite test with GPU
+echo "[test-all]: running Delite Cuda tests"
+sbt -Dtests.threads=1 -Dtests.targets=cuda "; project delite-test; test"
+
 # all Forge DSL tests
 echo "[test-all]: running Forge DSL tests"
 
@@ -40,7 +50,7 @@ do
     $FORGE_HOME/bin/update ${runners[$i]} $dsl 
     cd published/$dsl/
     echo "[test-all]: running $dsl tests (1 thread)"
-    sbt -Dtests.threads=1 -Dtests.targets=scala,cpp "; project $dsl-tests; test"
+    sbt -Dtests.threads=1 -Dtests.targets=scala,cpp,cuda "; project $dsl-tests; test"
     echo "[test-all]: running $dsl tests (8 threads)"
     sbt -Dtests.threads=8 -Dtests.targets=scala,cpp "; project $dsl-tests; test"
     popd
